Update Code Operasi

+++ Indomay/src/Membership.java
@@ -8,24 +8,25 @@ public class Membership {
      * - Cek promo (benefit membership)
      */
 
-    private int poin;
+    private double poin;
 
-    public Membership(){}
+    public Membership(){
+    }
 
     public Membership(int poin){
         this.poin = poin;
     }
 
-    public int getPoin() {
+    public double getPoin() {
         return poin;
     }
 
-    public void setPoin(int poin) {
+    public void setPoin(double poin) {
         this.poin = poin;
     }
 
     @Override
     public String toString(){
-        return String.format("%d", poin); 
+        return String.format("%.2f", poin); 
     }
 }
